+++ src/main/java/com/example/students/controllers/RegisterController.java
@@ -10,7 +10,6 @@ import jakarta.servlet.http.HttpServletRequest;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -27,42 +26,49 @@ public class RegisterController {
     private final AddressService addressService;
     private final SchoolService schoolService;
     private final NavbarService navbarService;
+    private final CookieService cookieService;
+    private final SessionService sessionService;
 
 
     @GetMapping
     public ModelAndView getRegister(HttpServletRequest request) {
+        if (cookieService.isSessionPresent(request.getCookies())) {
+            return new ModelAndView("register").addObject("user", new User()).addObject("address", new Address()).addObject("school", new School()).addObject("navElements", navbarService.getNavbar(cookieService.getValue(request.getCookies()),sessionService));
 
-        return new ModelAndView("register").addObject("user", new User()).addObject("address", new Address()).addObject("school", new School()).addObject("navElements", navbarService.getNavbar(request));
+        }
+        throw new RuntimeException();
     }
 
     @PostMapping
     @Transactional(rollbackFor = InvalidCredentialsException.class)
-    public String postRegister(@ModelAttribute User user, @ModelAttribute Address address, @ModelAttribute School school, Model model) {
+    public String postRegister(@ModelAttribute User user, @ModelAttribute Address address, @ModelAttribute School school, HttpServletRequest request) {
+        if (cookieService.isSessionPresent(request.getCookies())) {
+            //check if data is valid
+            try {
+                if (userService.checkEmailAndUsername(user)) {
+                    throw new InvalidCredentialsException();
+                }
+                user.hashPassword();
+                school.setAddress(address);
+                user.setSchool(school);
+                Principal principal = new Principal();
+                principal.setUser(user);
 
-        //check if data is valid
-        try {
-            if (userService.checkEmailAndUsername(user)) {
-                throw new InvalidCredentialsException();
+                addressService.saveAddress(address);
+                schoolService.saveSchool(school);
+                userService.saveUser(user);
+
+                if (principalService.insertPrincipal(principal) != null) {
+                    return "redirect:/login";
+                }
+            } catch (Exception e) {
+                userService.deleteUser(user);
+                schoolService.deleteSchool(school);
+                addressService.deleteAddress(address);
+                throw e;
             }
-            user.hashPassword();
-            school.setAddress(address);
-            user.setSchool(school);
-            Principal principal = new Principal();
-            principal.setUser(user);
-
-            addressService.saveAddress(address);
-            schoolService.saveSchool(school);
-            userService.saveUser(user);
-
-            if (principalService.insertPrincipal(principal) != null) {
-                return "redirect:/login";
-            }
-        } catch (Exception e) {
-            userService.deleteUser(user);
-            schoolService.deleteSchool(school);
-            addressService.deleteAddress(address);
-            throw e;
+            return "/index";
         }
-        return "/index";
+        throw  new RuntimeException();
     }
 }
